requirements.txt
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
jmespath==1.0.1
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tzdata==2024.2
urllib3==1.26.15
